+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0] - 2023-02-22
+
+- add admin command to toggle debug logging
+- add automated clearing of target panel before warp
+- add buff showing elevation with control tips
+- add support for local players
+- allow elevator blocks to be stacked up to 5
+- finalize readme and journal entries
+- prevent repeated warps from held key
+- rename project
+- unlock crouch on warp in local games
+- update code to match csharp standards
+- update to jump/crouch controls vs look angles
+- wrap debug logging in disabled debug flag
+
 ## [0.1.0] - 2022-11-17
 
 - add elevator 'look' controls